soc-agent/requirements.txt
fastapi>=0.109.0
uvicorn>=0.35
httpx>=0.28.1
python-dotenv>=1.1.0
pydantic>=2.11.0
chromadb>=0.5.0
